+++ src/Skeletest/Prop.hs
@@ -6,6 +6,12 @@ module Skeletest.Prop (
   setConfidence,
   setVerifiedTermination,
   setTestLimit,
+
+  -- * Coverage
+  classify,
+  cover,
+  label,
+  collect,
 ) where
 
 import Skeletest.Prop.Internal